Day 3 - APi chuẩn với MapStruct

+++ src/main/java/com/tpnam/spring_boot/service/UserService.java
@@ -1,6 +1,8 @@
 package com.tpnam.spring_boot.service;
 
+import com.tpnam.spring_boot.dto.UserDTO;
 import com.tpnam.spring_boot.exception.UserNotFoundException;
+import com.tpnam.spring_boot.mapper.UserMapper;
 import com.tpnam.spring_boot.model.User;
 import com.tpnam.spring_boot.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -14,27 +16,39 @@ public class UserService {
 
     private final UserRepository userRepository;
 
+    private final UserMapper userMapper;
+
     public List<User> getAll() {
         return userRepository.findAll();
     }
 
-    public User getById(Long id) {
-        return userRepository.findById(id).orElseThrow(() -> new UserNotFoundException("User not found with id: " + id));
+    public UserDTO getById(Long id) {
+        User user = userRepository.findById(id)
+                .orElseThrow(() -> new UserNotFoundException("User not found with id " + id));
+
+        return userMapper.toDto(user);
     }
 
-    public User create(User user) {
-        return userRepository.save(user);
+    public UserDTO  create(UserDTO  userDTO) {
+        User user = userMapper.toEntity(userDTO);
+        User savedUser = userRepository.save(user);
+        return userMapper.toDto(savedUser);
     }
 
-    public User update(Long id, User user) {
-        User existing = getById(id);
-        existing.setName(user.getName());
-        existing.setEmail(user.getEmail());
-        existing.setAge(user.getAge());
-        return userRepository.save(existing);
+    public UserDTO  update(Long id, UserDTO  userDTO) {
+        User existing = userRepository.findById(id)
+                .orElseThrow(() -> new UserNotFoundException("User not found with id " + id));
+        existing.setName(userDTO.getName());
+        existing.setEmail(userDTO.getEmail());
+        existing.setAge(userDTO.getAge());
+        return userMapper.toDto(existing);
     }
 
     public void delete(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new UserNotFoundException("User not found with id " + id);
+        }
+
         userRepository.deleteById(id);
     }
 }
